Self-contained to workaround pip install
Installing protobuf with pip install does not generate a useable
package, because __init__.py does not exist in the top level google
directory. Since the license allows for redistribution, it seemed easier
to just bundle a working version already.

This also fixes a bug with an undefined variable in
generate_compile_commands_json.py.